Refactor to a setFeedback base method
